+++ src/app/components/form/account/registration/FormAccountRegistrationAge.vue
@@ -43,9 +43,6 @@
             />
           </PopoverContent>
         </Popover>
-        <FormDescription>
-          {{ t('explanation') }}
-        </FormDescription>
         <FormMessage />
       </FormItem>
     </FormField>
@@ -66,21 +63,14 @@ import { useForm } from 'vee-validate'
 import { z } from 'zod'
 
 import { cn } from '@/utils/shadcn'
-import { useUpdateAccountBirthDateMutation } from '~~/gql/documents/mutations/account/accountBirthDateUpdate'
 
 const emit = defineEmits<{
-  submit: []
-  success: []
+  success: [string]
 }>()
-
-// api data
-const updateAccountBirthDateMutation = useUpdateAccountBirthDateMutation()
-const api = getApiData([updateAccountBirthDateMutation])
 
 // form
 const { t, locale } = useI18n()
 const templateForm = useTemplateRef('form')
-const modelError = defineModel<Error>('error')
 const submit = () => {
   templateForm.value?.dispatchEvent(
     new Event('submit', { bubbles: true, cancelable: true }),
@@ -89,7 +79,25 @@ const submit = () => {
 const { handleSubmit, setFieldValue, values } = useForm({
   validationSchema: toTypedSchema(
     z.object({
-      birthDate: z.string(),
+      birthDate: z
+        .string()
+        .nonempty()
+        .refine(
+          (value) => {
+            const birthDate = parseDate(value)
+            const todayDate = today(getLocalTimeZone())
+            const targetDate = todayDate.subtract({ years: 18 })
+
+            return birthDate.compare(targetDate) <= 0
+          },
+          {
+            params: {
+              i18n: {
+                key: 'globalFormErrorAge18',
+              },
+            },
+          },
+        ),
     }),
   ),
 })
@@ -98,34 +106,8 @@ const value = computed({
   set: (val) => val,
 })
 const onSubmit = handleSubmit(async (values) => {
-  const result = await updateAccountBirthDateMutation.executeMutation({
-    input: {
-      birthDate: values.birthDate,
-    },
-  })
-
-  if (result.error) return
-
-  if (!result.data) {
-    modelError.value = new Error(t('globalErrorNoData'))
-    return
-  }
-
-  emit('success')
+  emit('success', values.birthDate)
 })
-watch(
-  () => api.value.errors,
-  (current) => {
-    modelError.value = current?.length
-      ? new Error(
-          getCombinedErrorMessages(current, {
-            postgresP0002: t('postgresP0002'),
-            postgres23514: t('postgres23514'),
-          })[0],
-        )
-      : undefined
-  },
-)
 
 // calendar
 const dateFormatter = new DateFormatter(locale.value, {
@@ -140,15 +122,9 @@ defineExpose({
 
 <i18n lang="yaml">
 de:
-  explanation: Dein Geburtsdatum wird zur Berechnung deines Alters verwendet.
   label: Geburtsdatum
   placeholder: Wähle ein Datum
-  postgresP0002: Das Geburtsdatum konnte nicht gespeichert werden, weil kein Konto zugeordnet werden konnte.
-  postgres23514: Das Geburtsdatum kann nur ein Mal festgelegt werden. Kontaktiere den Support, wenn du Fragen hierzu hast.
 en:
-  explanation: Your date of birth is used to calculate your age.
   label: Date of birth
   placeholder: Pick a date
-  postgresP0002: The date of birth could not be saved because no associated account could be found.
-  postgres23514: The date of birth can only be set once. Contact support if you have any questions about this.
 </i18n>